Relative color syntax
https://csswg.sesse.net/css-color-5/#relative-colors

This implementation will be a lot cleaner when the color parsing
functions are unified, but this needs to get merged first:
https://chromium-review.googlesource.com/c/chromium/src/+/4598329

For handling inputs like:
oklab(from magenta calc(l * 2) a b);
The simplest method for handling variables in the calc parsing was to
create a map of CSSValueIDs to numerical values and pass it around,
finally substituting in CSSMathExpressionNode::ParseValue. This assumes
that all color channels are known at parse time, but there are currently
no tests in wpt that test with things like currentColor for input.

TODO:
- color() functions
- handle var() inputs
- allow approximate values to pass in color-valid-relative-color

Bug: 1447327
Change-Id: I865d41400f33fa5a8c0697aa3d7a61e51740d791