[CapturedMouseEvents] Export web platform tests
This patch exports existing WPT tests added in [1], using the directory
recently introduced in [2] and taking into account the recent renaming
to "captured mouse events".

It also adds a manual test that is verified to pass with [3] applied.
WPT does not provide support for more complex and automated tests,
so this is just a minimal verification. More advanced internal tests
with wider coverage will be added in [3].

[1] https://chromium-review.googlesource.com/c/chromium/src/+/4517372
[2] https://github.com/web-platform-tests/wpt/pull/40826
[3] https://chromium-review.googlesource.com/c/chromium/src/+/4549484

Bug: 1444712
Change-Id: I8ed455c406d6cba81cb4df0651509b6d6f7bd991