Update the common ancestor finding logic for mouseup with mousedown
The existing function doesn't not work for cases like when mousedown
and mouseup were happened in the same hierarchy but the mousedown target
was a different interactive element than the mouseup target. This
could be easily accomplished by setting `pointer-events: none` or
`display: none` to the `active` pseudo-class of the mousedown target.

This patch fixes this by when the logic tries to find the event target
tree of the mousedown target, it continues building up (by recursively looking
up the parents) the tree instead of stopping at the first interactive element.

Differential Revision: https://phabricator.services.mozilla.com/D178844

bugzilla-url: https://bugzilla.mozilla.org/show_bug.cgi?id=1792110
gecko-commit: d492bd7d93d7b61792b9a149f2fb9ae94ffc3c61
gecko-reviewers: edgar