Synthetic gestures only dispatch to visible widget
This UAF is caused by a synthetic pointer being routed to browser UI. In
this case, it starts dragging a tab which starts a nested message loop
Further events and gestures are processed in this nested loop and
cleaned up. When the message loop returns the stack contains the cleaned
up pointers.

Synthetic gestures shouldn't be able to target UI outside the web
contents area. The event location is intersected with the web contents'
RenderWidget's view bounds to prevent this [1]. However, the bounds will
be inaccurate if the widget is in a background tab; it won't receive
resizes until it's foregrounded (it's also bad that we can dispatch
events to a different tab).

This CL fixes the issue by ensuring events are dispatched only to a
foregrounded widget. If a synthetic gesture is started while the widget
is in a background tab, its start is deferred until it comes into the
foreground. If the widget is backgrounded while a gestuere is in
progress, the gesture is aborted.

Note: we don't do this for DevTools injected events as those skip event
routing and go straight to the injecting renderer. The comment in [2]
makes me think this is a common use case.

[1] https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:content/browser/renderer_host/input/synthetic_gesture_target_base.cc;l=155;drc=ac872e771ce001fef191848bab4167d60dfda403
[2] https://source.chromium.org/chromium/chromium/src/+/refs/heads/main:content/browser/renderer_host/input/synthetic_gesture_target_aura.cc;l=140;drc=ac872e771ce001fef191848bab4167d60dfda403

Bug: 1444597
Change-Id: I2955ce60357f7f03e62f44fd1497bd4ea598f660