Add report-lite to stash reports from PAA test for protected audience
Create a lite version report script to receive message during protected
audience test cases. It only supports passing reports via parameters and
the reports are stored by uuid.

For the successful test cases, a sendReportTo()/reportAdAuctionWin() is triggered after
calling contributeToHistogram(). It will send a report of
the worklet's name to the report-lite server as the value in the stash.

For the failure test cases, a sendReportTo()/reportAdAuctionWin() is triggered when an
exception is caught. The report format is same as above.

At the end of each test case, there is a observe function which will
try to get the stashed reports by the uuid to check if the reports are
sent and stashed successfully to verify contributeToHistogram() is
called with or without errors.

In the report-lite server, the GET is set with a timeout of 15 seconds.

Bug: 1456401
Change-Id: I7382203a8a8a4da14667d9e7eddeef16c89ffe73